design patterns: adapter
