+++ examples/demo_colibri.py
@@ -50,7 +50,7 @@ else:
 # %%
 # Load dataset
 # -----------------------------------------------
-from colibri.data.datasets import Dataset
+from colibri.data.datasets_ import Dataset
 
 dataset_path = 'cifar10'
 keys = ''
